Add JasmineGJSBootstreap, EndlessConsoleReporter and EndlessTimer
This enables Jasmine to run on top of Gjs.

EndlessConsoleReporter is a replacement for the default console
reporter which prints out more details about which specs are
actually running and also prints out stacktraces of failed
specs.

EndlessTimer is an implementation of the Timer interface using
GLib timers.

An 'eos-jasmine-run' script is provided for Automake integration.

In order to integrate with Automake, this script runs only one file at
a time, given on the command line. It is meant to be used when the
eos-jasmine repo is included as a submodule in another repo.

The Makefile.am must have the following modifications:

- Make sure EXTRA_DIST, TEST_EXTENSIONS, and AM_JS_LOG_FLAGS are defined
  (if necessary as empty variables)
- Define the variable JASMINE_SUBMODULE_PATH to the path where the
  Jasmine submodule is in your source tree (use $(top_srcdir) to ensure
  that VPATH builds work
- Add 'include path/to/jasmine/Makefile-jasmine.am.inc' to include the
  custom Jasmine Makefile code

[endlessm/eos-jasmine#1]