Improve readability of test-suite.log.
When running make check in a repo with a failing test that uses this
reporter, the output in test-suite.log is quite noisy, complaining about
an exception being raised outside of the test suite. We exit with a
non-zero exit code instead so that test-suite.log still considers the
failed test to have failed but less noise is produced.

[endlessm/eos-sdk#2017]